Add workshop solution folder Rename current solution folder to 'stretch-goal-solution'
