+++ stretch-goal-solution/public/index.html
@@ -1,0 +1,17 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+
+  <link rel="stylesheet" href="./style.css">
+</head>
+
+<body>
+  <div id="root"></div>
+
+  <script src="../src/index.js"></script>
+</body>
+
+</html>
