+++ stretch-goal-solution/public/style.css
@@ -1,0 +1,5 @@
+body {
+  background: #e0e2de;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
+    Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
+}
