+++ stretch-goal-solution/src/components/app.js
@@ -1,0 +1,6 @@
+import React from "react";
+import Card from "./card";
+
+export default function App() {
+  return <Card />;
+}
