+++ stretch-goal-solution/src/components/card/card_image/style.css
