+++ stretch-goal-solution/src/components/card/style.css
