+++ stretch-goal-solution/src/components/card/user_info/style.css
