+++ stretch-goal-solution/src/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import "sanitize.css/sanitize.css";
+
+import App from "./components/app";
+
+ReactDOM.render(<App />, document.getElementById("root"));
