+++ stretch-goal-solution/src/utils/data_helpers.js
