+++ workshop/src/components/app.js
@@ -1,9 +1,8 @@
 import React from "react";
-import UserHeader from "./UserHeader"
 
 export default function App() {
-
-  return (
-    <UserHeader />
-  )
+  const [placeholder, setPlaceholder] = React.useState(
+    "Let's get started! You can delete this state as you won't need it going forward."
+  );
+  return <div>{placeholder}</div>;
 }
